requirements.txt
ccxt==4.3.89
pandas==2.2.2
numpy
ta==0.11.0
lightgbm==4.5.0
scikit-learn==1.5.1
vectorbt==0.26.1
numba
bottleneck==1.3.8
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
pyyaml==6.0.2
psutil>=5.9.0
websocket-client>=1.6.0



